reorganization finished
