+++ web/constants.py
@@ -1,0 +1,1 @@
+MB_UPLOAD_LIMIT = 10
