+++ web/samples.py
@@ -1,0 +1,85 @@
+import datetime
+import os
+import re
+import uuid
+import pathlib
+
+from flask import jsonify
+# from flask import abort, flash, jsonify, redirect, render_template, request, session, url_for
+# from flask_login import current_user, login_user, logout_user
+
+from models import Sample, db, User
+from utils import add_sample_to_db, allowed_file, check_video, create_thumbnail, reencode_video
+import utils
+
+from sqlalchemy import func
+from werkzeug.utils import secure_filename
+
+from constants import MB_UPLOAD_LIMIT
+    
+def edit_sample(filename,stored_as,thumbnail,uploader,source_id,reencode):
+    if reencode:
+        utils.reencode_video(stored_as)
+
+
+    add_sample_to_db(
+        filename,
+        stored_as,
+        datetime.datetime.now(datetime.UTC),
+        str(thumbnail),
+        uploader,
+        source_id,
+    )
+
+def upload(file, sample_ids):
+    if file and allowed_file(file.filename):
+        original_filename = secure_filename(file.filename)
+
+        timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S%f")[:-3]
+        random_id = uuid.uuid4().hex[:6]
+        
+        stored_as = f"{os.path.splitext(original_filename)[0]}_{timestamp}_{random_id}.mp4"
+        stored_as = re.sub(r"[^\w\s.-]", "", stored_as)
+
+        # current_time = int(time.time() * 100)
+        # filename = secure_filename(file.filename + str(current_time))
+
+        # Make sure the directories actually exist
+        pathlib.Path("static/media/samps").mkdir(parents=True,exist_ok=True)
+        pathlib.Path("static/media/thumbs").mkdir(parents=True,exist_ok=True)
+
+        upload_path = os.path.join("static/media/samps", stored_as)
+        file.save(upload_path)
+
+        if not check_video(upload_path):
+            os.remove(upload_path)
+            raise Exception("There is an error one of your files. Please make sure it is a valid .mp4 file.")
+        
+        if os.path.getsize(upload_path) > MB_UPLOAD_LIMIT * 1000 * 1000:
+            os.remove(upload_path)
+            raise Exception("Content is too large")
+
+        create_thumbnail(upload_path, f"static/media/thumbs/{timestamp}.png")
+
+        sample_id = str(uuid.uuid4())
+        sample_ids.append(sample_id)
+
+    return (sample_id, original_filename, timestamp, stored_as)
+
+def delete_sample(sample_id):
+    sample = Sample.query.get(sample_id)
+    if sample:
+        warnings = []
+        try:
+            os.remove(os.path.join("static/media/thumbs", sample.thumbnail_filename))
+        except FileNotFoundError as _:
+            warnings.append("Thumbnail file wasn't found, couldn't be deleted")
+        try:
+            os.remove(os.path.join("static/media/samps", sample.stored_as))
+        except FileNotFoundError as _:
+            warnings.append("Sample file wasn't found, couldn't be deleted")
+        db.session.delete(sample)
+        db.session.commit()
+        return jsonify({"message": "Sample deleted successfully.", "warnings": warnings})
+    return jsonify({"message": "There was an error deleting the sample."})
+
